requirements.txt
Flask==3.0.1
gunicorn==21.2.0
tensorflow-cpu==2.16.0rc0
# tensorflow==2.16.1
numpy==1.26.3
Pillow==10.2.0